requirements.txt
selenium>=4.15.0
undetected-chromedriver>=3.5.4
python-telegram-bot>=20.7
pydantic>=2.5.0
python-dotenv>=1.0.0

